requirements.txt
aylien-news-api
flask
python-dotenv
autopep8
flask-cors
aylien-apiclient